Moved FEMNIST model
